Move depth measures and stats to exploratory
