+++ docs/modules/exploratory/.gitignore
@@ -1,0 +1,1 @@
+/autosummary/
