+++ skfda/exploratory/stats/__init__.py
@@ -1,0 +1,1 @@
+from ._stats import mean, var, gmean, cov
