+++ skfda/exploratory/stats/_stats.py
@@ -1,0 +1,68 @@
+"""Functional data statistics.
+"""
+
+
+def mean(fdata):
+    """Compute the mean of all the samples in a FData object.
+
+    Computes the mean of all the samples in a FDataGrid or FDataBasis object.
+
+    Args:
+        fdata(FDataGrid or FDataBasis): Object containing all the samples
+        whose mean
+            is wanted.
+
+    Returns:
+        FDataGrid or FDataBasis: A FDataGrid or FDataBasis object with just
+        one sample representing the mean of all the samples in the original
+        object.
+
+    """
+    return fdata.mean()
+
+
+def var(fdatagrid):
+    """Compute the variance of a set of samples in a FDataGrid object.
+
+    Args:
+        fdatagrid (FDataGrid): Object containing all the set of samples
+        whose variance is desired.
+
+    Returns:
+        FDataGrid: A FDataGrid object with just one sample representing the
+        variance of all the samples in the original FDataGrid object.
+
+    """
+    return fdatagrid.var()
+
+
+def gmean(fdatagrid):
+    """Compute the geometric mean of all the samples in a FDataGrid object.
+
+    Args:
+        fdatagrid (FDataGrid): Object containing all the samples whose
+            geometric mean is wanted.
+
+    Returns:
+        FDataGrid: A FDataGrid object with just one sample representing the
+        geometric mean of all the samples in the original FDataGrid object.
+
+    """
+    return fdatagrid.gmean()
+
+
+def cov(fdatagrid):
+    """Compute the covariance.
+
+    Calculates the covariance matrix representing the covariance of the
+    functional samples at the observation points.
+
+    Args:
+        fdatagrid (FDataGrid): Object containing different samples of a
+            functional variable.
+
+    Returns:
+        numpy.darray: Matrix of covariances.
+
+    """
+    return fdatagrid.cov()
